--- packages/client-react-streaming/src/ExperimentalReact/index.ts
@@ -1,1 +1,0 @@
-export { ExperimentalReactDataTransport } from "./ExperimentalReactDataTransport.js";
